Move cli tests to main, add collector tests
